Feature: User can now search images online and use them as face swap targets.
Implement UrlImageDownloader and Pexels API integration.
- Added UrlImageDownloader class for downloading images from URLs using libcurl.
- Integrated Pexels API for image searching and downloading.
- Enhanced Webcam class to support additional image formats.
- Updated Application class to manage target images and Pexels API client initialization.
- Modified Config class to load environment variables from a .env file.
- Improved UI to include a search feature for Pexels images with thumbnail display.
- Added tests for UrlImageDownloader to validate constructor and error handling.

+++ include/LinuxFace/codec.h
@@ -717,14 +717,14 @@ class YUV422 : public Decoder
             return false;
         }
 
-        const size_t expectedSize = width_ * height_ * 2;
+        const size_t expectedSize = static_cast<size_t>(width_ * height_ * 2u);
         if (srcImage.size() < expectedSize)
         {
             common::logError("YUV422::decode - Input buffer too small: %zu < %zu", srcImage.size(), expectedSize);
             return false;
         }
 
-        const size_t rgbSize = width_ * height_ * 3;
+        const size_t rgbSize = static_cast<size_t>(width_ * height_ * 3u);
         if (outImage.size() != rgbSize)
         {
             outImage.resize(rgbSize);
@@ -734,7 +734,7 @@ class YUV422 : public Decoder
         outImage.info = srcImage.info;
         outImage.info.width = width_;
         outImage.info.height = height_;
-        outImage.info.pixelSizeBytes = 3;
+        outImage.info.pixelSizeBytes = 3u;
         outImage.info.TJPixelFormat = TJPF_RGB;
 
         return decodeYUV(srcImage, outImage, pxConvert);
